post_service/requirements.txt
flask>=2.2.5
requests==2.26.0
grpcio==1.48.2
grpcio-tools==1.48.2
grpcio-health-checking
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2